third_party/WebKit/LayoutTests/fast/css/atapply/at-apply-nested.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
#e {
    width: 100px;
    height: 200px;
    --bar: {
        background-color: green;
    };
    --foo: {
        background-color: red;
        @apply --bar;
    };
    @apply --foo;
}
</style>
<div id=e>
</div>
<script>
test(function(){
    assert_equals(getComputedStyle(e).backgroundColor, "rgb(0, 128, 0)");
    assert_equals(getComputedStyle(e).width, "100px");
    assert_equals(getComputedStyle(e).height, "200px");
    assert_equals(getComputedStyle(e).getPropertyValue("--foo"), " { background-color: red;  background-color: green;  }");
}, "@apply rules can be used inside variables");
</script>
